refactor: hange name of currently named cmdWrapper.js file
Name has been changed successgully

Closes #20
