--- src/little-smart.2018-07-10.private-key.pem
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEpAIBAAKCAQEAsSvCqToKP5/pzNQ640F539aT63p+6kcjJ6jlzwQKKsKD0T2j
-q/xEbmo2eLBjv2R944p6MhXq+HWJd5lTDTEMXPZ1pGwSutNcFBVHKh7cdUduG03G
-HmzEkRQk4veZLuWMgOFvsq2tJoRjnhhMxa/jQPQW5q5Ukv+H9gnbnn3aH6GqvUMV
-2w1fpfjMG28cjpwEUYcUvzMPKEmSLfpTmU9LqbR2eymZYYRKuZelr4HlAufg2lYd
-FXaIRRUAKdwdiij+V1EMfAPjdSYJUDuUFrQplXnveT+btkXgcWS0DxXlwOt/1JD+
-eUw8Ak+SK6Vsrj+EO2q+YDpbnpLRtcxtdCSbMwIDAQABAoIBAQCoI5OV7pMTVrwX
-u8I3MkhrecTf0VzpEaH3XX7bPEY51mgrB4hxC/QVpY8UcrH3TXN7okvdhJd+PG29
-pzD+b/H/t5Zl+Hq9+yCDnr3u52LpmmSQXPsMnMTfunJXRAuISilRlqDevZx51WKF
-AyK46feOCRhCzGdPKlrp0OkwPNfo83uC3Xe/nDmYeCrFt2D3jSugnvu9mtTqlVQF
-qQOK1noUw/RnP2bv/HZoI1E3jur9IxWbf/UrBEfKXHLCMIyg3L2rBl2jYX6QEuSR
-v6sb1EwT7q1ils3SFKTVFdvnwsG9FfpX6uKo5ivpJIR2IkVAgMd8n6H1ixXAhF/Y
-ZPLQ+6FBAoGBAOLH4JgsvyxidQkdDZ3lkBcvGdCSsoh6HY3pxq/daKWHmPFvyehm
-+ylpPn23ZOWbDCKFABOgfIfR3If7uLInBcrDfA+u8KsUYqIY+XpTHvEIl8errYto
-YXudGQWfbtOB2lfjJkbm7HWYUf635lUvCn+vyuB90fAu/rWp7ZClpYPpAoGBAMf/
-jjVy0iNwWwhdlXHBxE+ZSWfypzTQq+Ho+FCGbwzSpLn/Bsrx7WtiH1QlnpNzU0cj
-pBtbjNEfFpC3ofvWWPZ2uS1MDLV7aaNsH5ahMR6iie99mo62Ozra76BrJZ/e+9qW
-RENJb1BWIhH78UsVe52jncPh/fMJIVp5mbM+S0C7AoGASZkenaRgArKNUwDXy3eW
-FMAXe6YwRLSE5OqKrcaIULg5/ZkPCNprBpx+SGR9XTq9bHKHfNcnsxh7nrIl1cNu
-UMiyeZCH0K3T+NFsHbPeJdFpGMbxkn20rV5loXqFiCTrTHyOjpIhCWT+r1DVcq3K
-lXfiuop62J0X0rMsQzZPCgkCgYA3W4klSi7YGFVaL20aG6QW1Dr5RrT5bXjmAvIY
-d91tEzD9OpXKvsQUoQhC5lbKQXWfj0HxQJntmElKrUlsx5vAPIRtAUvTP+Gxlk/4
-BY3vCNZdL8sEDXS/ibIiGQZFBhd4HXevmDfzaKPtjmrveXbJ32vfuEHub2yH+niM
-oIV72QKBgQCsp/YTtl6tHVTcZdO2S3x9vS9mjqnZ7gpaI2WW1xYMLiII8acihrbQ
-mcFH9omhLjXSv1meB1jNhLkMlL/IEBIMUqWfpKgI/65qySikkUs1M72FYXP9qaOC
-RojQBiDFlQ7yC2f+aLitN2c/I2Os9G/I4nslxyxvkDc5DUFfo78E+A==
------END RSA PRIVATE KEY-----
